requirements.txt
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
django-leaflet==0.26.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
requests==2.23.0
whitenoise==5.0.1


